+++ src/app/(authorized)/Global.css
@@ -1,0 +1,62 @@
+*,
+html,
+body {
+  margin: 0;
+  padding: 0;
+  border: 0;
+  font: inherit;
+  font-size: 10px;
+  box-sizing: border-box;
+  line-height: 1.5;
+  font-family: sans-serif;
+}
+body {
+  background: url("https://cdn.architextures.org/textures/20/08/buff-common-5f2fdc270073c-1200.jpg");
+  background-repeat: repeat;
+  background-size: 100px 100px;
+}
+
+ol,
+ul {
+  list-style: none;
+}
+
+table {
+  border-collapse: collapse;
+  border-spacing: 0;
+}
+
+@media (max-width: 800px) {
+  html {
+    font-size: 8px;
+  }
+}
+@media (max-width: 500px) {
+  html {
+    font-size: 6px;
+  }
+}
+
+@media (max-height: 500px) {
+  html {
+    font-size: 5px;
+  }
+}
+
+@media (max-height: 400px) {
+  html {
+    font-size: 4px;
+  }
+}
+
+@media (max-height: 300px) {
+  html {
+    font-size: 2.5px;
+  }
+}
+
+@media (max-height: 250px) {
+  html {
+    font-size: 2px;
+  }
+}
