+++ src/app/(authorized)/blog/[id]/page.css
