+++ src/app/(authorized)/blog/[id]/page.jsx
