+++ src/app/(authorized)/layout.jsx
@@ -1,4 +1,7 @@
 import "./Global.css";
+import Header from "@/Components/Header/Header";
+import Footer from "@/Components/Footer/Footer";
+
 export const metadata = {
   title: "Create Next App",
   description: "Generated by create next app",
@@ -9,7 +12,9 @@ export default function RootLayout({ children }) {
     <html lang="en">
       <body>
         <div>
+          <Header />
             {children}
+          <Footer />
         </div>
       </body>
     </html>
