+++ src/app/(authorized)/products/[id]/page.css
