+++ src/app/(authorized)/products/[id]/page.jsx
