+++ src/app/(authorized)/products/page.css
