+++ src/app/(authorized)/products/page.jsx
