+++ src/app/(unauthorized)/login/login.css
@@ -62,9 +62,10 @@
   background-color: #0056b3;
 }
 
-p {
+.login-form > p {
   text-align: center;
   color: red;
+  font-size: 1.5rem;
 }
 
 .video-container {
