--- src/app/temporary/Temporary.css
@@ -1,29 +1,0 @@
-.Temporary {
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-table {
-  width: 70%;
-  border-collapse: collapse;
-  margin: 2rem 0;
-  font-size: 1.5rem;
-}
-
-th,
-td {
-  padding: 1rem;
-  text-align: left;
-  border: 1px solid #9c9999;
-}
-
-thead {
-  background-color: #c8c5c5;
-}
-
-tbody tr:nth-child(even) {
-  background-color: #f9f9f9;
-}
