+++ src/middleware.js
@@ -1,0 +1,15 @@
+import { NextResponse } from 'next/server';
+
+export function middleware(request) {
+  const token = request.cookies.get('accessToken');
+  
+  if (!token && request.nextUrl.pathname !== '/login') {
+    return NextResponse.redirect(new URL('/login', request.url));
+  }
+
+  return NextResponse.next();
+}
+
+export const config = {
+  matcher: ['/((?!login|_next|public).*)'],
+};
